Extract _get_manage helper from cron.py to paths.py
Move _get_manage() and _manage_cache from cron.py to paths.py so
other modules can use it without depending on the cron subgroup.
Update imports in cron.py and list_cmd.py accordingly.
